+++ Anarchy/Commands/Command.cs
@@ -1,7 +1,9 @@
-﻿namespace Discord.Commands
+﻿using Discord.Gateway;
+
+namespace Discord.Commands
 {
     public abstract class Command
     {
-        public abstract void Execute(string[] args, Message message);
+        public abstract void Execute(DiscordSocketClient client, string[] args, Message message);
     }
 }