+++ Anarchy/Commands/CommandAttribute.cs
@@ -6,10 +6,12 @@ namespace Discord.Commands
     public class CommandAttribute : Attribute
     {
         public string Command { get; private set; }
+        public string Description { get; private set; }
 
-        public CommandAttribute(string cmd)
+        public CommandAttribute(string cmd, string description = null)
         {
             Command = cmd;
+            Description = description;
         }
     }
 }