+++ Anarchy/REST/Guild/Models/SocketGuildEventArgs.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace Discord.Gateway
+{
+    public class SocketGuildEventArgs : EventArgs
+    {
+        public SocketGuild Guild { get; private set; }
+
+
+        public SocketGuildEventArgs(SocketGuild guild)
+        {
+            Guild = guild;
+        }
+    }
+}